Update CHANGELOG.
Renames cmr => cmr-class.
Renames granule-extract-cmr => cmr.
Updates imports and tests.
