+++ packages/cmrjs/index.js
@@ -6,7 +6,7 @@ const {
   ingestConcept,
   deleteConcept,
   CMR
-} = require('./cmr');
+} = require('./cmr-class');
 const {
   ValidationError,
   updateToken,
@@ -20,7 +20,7 @@ const {
   getCmrFiles,
   publish,
   updateMetadata
-} = require('./granule-extract-cmr');
+} = require('./cmr');
 
 
 /**
